+++ src/pages/detail/FontDetail.jsx
@@ -3,9 +3,9 @@ import styled from 'styled-components'
 import { FontDeleteButton } from '@/features/delete-font/ui/FontDeleteButton'
 import { FontEditButton } from '@/features/edit-font/ui/FontEditButton'
 import { SectionLayout } from '@/shared/ui/SectionLayout'
-import { FontDetailContent } from '@/widgets/font-detail/ui/FontDetailContent'
-import { FontDetailHeader } from '@/widgets/font-detail/ui/FontDetailHeader'
-import { RecommendFontList } from '@/widgets/font-list/ui/RecommendFontList'
+import { RecommendFontList } from '@/widgets/font-list/RecommendFontList'
+import { FontDetailHeader } from '@/widgets/header/FontDetailHeader'
+import { FontDetailContent } from '@/widgets/section/FontDetailContent'
 
 const dummyFont = {
   profile: {
