+++ src/pages/my-font/MyCustomFont.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { SearchStoreProvider } from '@/features/search/model/searchProvider.provider'
 import { SearchBar } from '@/features/search/ui/SearchBar'
 import { SectionLayout } from '@/shared/ui/SectionLayout'
-import { MyCustomFontList } from '@/widgets/font-list/ui/MyCustomFontList'
+import { MyCustomFontList } from '@/widgets/font-list/MyCustomFontList'
 
 export const MyCustomFont = () => {
   return (
