+++ src/widgets/card/NavigationCard.jsx
