+++ src/widgets/font-list/item/FontItemWithProfile.jsx
@@ -1,17 +1,13 @@
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { BookmarkButton } from '@/features/bookmark/ui/BookmarkButton'
-import { DownloadButton } from '@/features/download/ui/DownloadButton'
-import { CaretRightIcon } from '@/shared/ui/icons/NonPropIcons'
+import { BookmarkButton } from '@/features/bookmark-font/ui/BookmarkButton'
+import { DownloadButton } from '@/features/download-font/ui/DownloadButton'
+import { FontProfile } from '@/shared/ui/FontProfile'
 
-export const FontItemWithoutProfile = ({ font }) => (
+export const FontItemWithProfile = ({ font }) => (
   <LinkContainer to={`/detail/${font.id}`}>
-    <FontInformation>
-      <FontName>{font.fontName}</FontName>
-      <CaretRightIcon />
-    </FontInformation>
-
+    <FontProfile size="md" fontName={font.fontName} authorName={font.authorName} />
     <PreviewText>{font.preview}</PreviewText>
 
     <ActionButtonGroup>
@@ -25,22 +21,15 @@ const LinkContainer = styled(Link)`
   ${({ theme }) => `
     ${theme.flexBox('column', undefined, undefined, '3xl')}
     ${theme.padding('lg', 'md', 'sm')}
-    ${theme.borderRadius('md')}
     ${theme.border('divider')}
+    ${theme.borderRadius('md')}
   `}
 `
 
-const FontInformation = styled.div`
-  ${({ theme }) => theme.flexBox('row', 'center', 'space-between')};
-`
-
-const FontName = styled.h6`
-  ${({ theme }) => theme.font(400)};
-`
-
 const PreviewText = styled.p`
-  ${({ theme }) => theme.font(700, theme.colors.grey[600])};
+  ${({ theme }) => theme.font(500, theme.colors.grey[600])};
   flex-grow: 1;
+  height: fit-content;
 `
 
 const ActionButtonGroup = styled.div`
