+++ src/widgets/section/FontDetailContent.jsx
