+++ src/widgets/table/FontProgressTable.jsx
